 bull_put_spread_algorithm/indicators.py
"""
indicators.py
--------------
Technical indicator computation for pre-trade screening.

Author: Paul K. Mwangi
"""

# Imports and basic configuration
# Purpose: group and document all imports and top-level configuration used across
# the Bull Put Spread strategy.

# -------------------------
# Standard library imports
# -------------------------
import re
import threading
import logging
import traceback
import os, json
import time
from zoneinfo import ZoneInfo
from datetime import date, datetime, timedelta, timezone

# -------------------------
# Numerical & statistics
# -------------------------
# pandas / numpy are used for data manipulation and numeric arrays. scipy provides
# probability functions and solvers used by pricing/IV calculations.
import pandas as pd
import numpy as np
from scipy.stats import norm
from scipy.optimize import brentq

# -------------------------
# Technical indicators
# -------------------------
# `ta` is a light-weight technical analysis library used for indicators such as
# MACD and ADX. These indicators provide part of the entry/exit signal logic.
from ta.trend import MACD, ADXIndicator

# -------------------------
# Typing and developer helpers
# -------------------------
# Used to provide type hints which improves readability and helps code review.
from typing import TypedDict, Optional, Dict, Any, List, Tuple
from collections import defaultdict
from bisect import bisect_left

# -------------------------
# Market data helper (optional)
# -------------------------
import yfinance as yf

# -------------------------
# Alpaca (data + trading)
# -------------------------
# Alpaca packages provide both historical & live data access and trading client
# objects. They are central for live/paper trading and option chain retrieval.
from alpaca.data.timeframe import TimeFrame
from alpaca.data.historical.option import OptionHistoricalDataClient
from alpaca.data.historical.stock import StockHistoricalDataClient
from alpaca.data.requests import StockBarsRequest, OptionLatestQuoteRequest
from alpaca.trading.client import TradingClient
from alpaca.trading.stream import TradingStream
from alpaca.data.live.stock import StockDataStream
from alpaca.trading.requests import (
    GetOptionContractsRequest,
    MarketOrderRequest,
    GetAssetsRequest,
    LimitOrderRequest,
    GetCalendarRequest,
    OptionLegRequest
)
from alpaca.trading.enums import (
    AssetStatus,
    OrderSide,
    OrderClass,
    TimeInForce,
    ContractType,
    AssetClass,
    PositionSide
)
def calculate_implied_volatility(self, option_price, S, K, T, r, option_type):
    """
    Estimate the implied volatility using the Black-Scholes model via Brent's method.
    Handles options close to intrinsic value gracefully.
    """

    return


def calculate_greeks(self, option_price, strike_price, expiration, underlying_price,
                     risk_free_rate, option_type, IV=None):
    """
    Calculate Black-Scholes Greeks (Delta, Gamma, Theta, Vega) with fallback handling.
    """

    return

    # ================================
    # (B) Technical Indicators
    # ================================


def calculate_rsi(self, series, period=14):
    """
    Calculate Relative Strength Index (RSI) using exponential smoothing.
    Args:
        series (pd.Series): Series of closing prices.
        period (int): Lookback period (default=14).

    Returns:
        pd.Series: RSI values scaled 0–100.
    """

    return


def calculate_moving_averages(self, df, short=20, long=50):
    """
    Compute short- and long-term Simple Moving Averages (SMA).

    Args:
        df (pd.DataFrame): Price data with 'close' column.
        short (int): Short-term window size.
        long (int): Long-term window size.

    Returns:
        pd.DataFrame: Original DataFrame with added SMA columns.
    """

    return df


def calculate_macd_features(self, df, fast=12, slow=26, signal=9):
    """
    Compute Moving Average Convergence Divergence (MACD).

    Args:
        df (pd.DataFrame): Price data with 'close' column.
        fast (int): Fast EMA window.
        slow (int): Slow EMA window.
        signal (int): Signal line EMA window.

    Returns:
        pd.DataFrame: DataFrame with MACD, Signal, and Histogram columns.
    """


def calculate_adx_features(self, df, period=14):
    """
    Compute Average Directional Index (ADX) to measure trend strength.

    Args:
        df (pd.DataFrame): Price data with 'high', 'low', 'close' columns.
        period (int): Lookback period for ADX.

    Returns:
        pd.DataFrame: DataFrame with added ADX column.
    """

    return df


def add_all_features(self, dataframe: pd.DataFrame) -> pd.DataFrame:
    """
    Compute and append all technical indicators used for signal generation.

    Args:
        dataframe (pd.DataFrame): OHLCV price data.

    Returns:
        pd.DataFrame: DataFrame enriched with:
                      - SMA(20), SMA(50)
                      - MACD, Signal, MACD Histogram
                      - ADX (trend strength)
                      - RSI (momentum)
                      - Volume SMA(20)
    """