 bull_put_spread_algorithm/requirements.txt
alpaca-py
pandas
numpy
scipy
ta
logging
os
json
